ai-agent/requirements.txt
openai>=1.0.0
redis>=5.0.0
confluent-kafka>=2.0.0
python-dotenv>=1.0.0
mysql-connector-python>=8.0.0
qdrant-client>=1.7.0
requests>=2.31.0
matplotlib>=3.8.0
tenacity>=8.2.0
python-json-logger>=2.0.7
aiomysql>=0.2.0
httpx>=0.27.0
cryptography>=42.0.0
